make separate query for node list component
