+++ src/routes/staking/staking.tsx
@@ -1,7 +1,7 @@
 import { useAppState } from "../../contexts/app-state/app-state-context";
 import { Error, Tick } from "../../components/icons";
 import { Link, useRouteMatch } from "react-router-dom";
-import { NodeList, NodeListItemProps } from "./node-list";
+import { NodeList } from "./node-list";
 import { Trans, useTranslation } from "react-i18next";
 
 import { BigNumber } from "../../lib/bignumber";
@@ -9,9 +9,7 @@ import { BulletHeader } from "../../components/bullet-header";
 import { Callout } from "../../components/callout";
 import { ConnectToVega } from "./connect-to-vega";
 import { Links } from "../../config";
-import React from "react";
 import { Staking as StakingQueryResult } from "./__generated__/Staking";
-import { useVegaUser } from "../../hooks/use-vega-user";
 import { useWeb3 } from "../../contexts/web3-context/web3-context";
 import { formatNumber } from "../../lib/format-number";
 
@@ -170,57 +168,5 @@ export const StakingStepSelectNode = ({
 }: {
   data?: StakingQueryResult;
 }) => {
-  const { t } = useTranslation();
-  const { currVegaKey } = useVegaUser();
-
-  const nodes = React.useMemo<NodeListItemProps[]>(() => {
-    if (!data?.nodes) return [];
-
-    const nodesWithPercentages = data.nodes.map((node) => {
-      const stakedTotal = new BigNumber(
-        data?.nodeData?.stakedTotalFormatted || 0
-      );
-      const stakedOnNode = new BigNumber(node.stakedTotalFormatted);
-      const stakedTotalPercentage =
-        stakedTotal.isEqualTo(0) || stakedOnNode.isEqualTo(0)
-          ? "-"
-          : stakedOnNode.dividedBy(stakedTotal).times(100).dp(2).toString() +
-            "%";
-
-      const userStake = data.party?.delegations?.length
-        ? data.party?.delegations
-            ?.filter((d) => d.node.id === node.id)
-            ?.filter((d) => d.epoch === Number(data.epoch.id))
-            .reduce((sum, d) => {
-              const value = new BigNumber(d.amountFormatted);
-              return sum.plus(value);
-            }, new BigNumber(0))
-        : new BigNumber(0);
-
-      const userStakePercentage =
-        userStake.isEqualTo(0) || stakedOnNode.isEqualTo(0)
-          ? "-"
-          : userStake.dividedBy(stakedOnNode).times(100).dp(2).toString() + "%";
-
-      return {
-        id: node.id,
-        name: node.name,
-        pubkey: node.pubkey,
-        stakedTotal,
-        stakedOnNode,
-        stakedTotalPercentage,
-        userStake,
-        userStakePercentage,
-        epoch: data.epoch,
-      };
-    });
-
-    return nodesWithPercentages;
-  }, [data]);
-
-  if (!currVegaKey) {
-    return <p className="text-muted">{t("connectVegaWallet")}</p>;
-  }
-
-  return <NodeList nodes={nodes} epoch={data?.epoch} />;
+  return <NodeList epoch={data?.epoch} party={data?.party} />;
 };
